src/domain/forum/aplication/use-cases/register-student.ts
import { Either, left, right } from '@/core/either'
import { Injectable } from '@nestjs/common'
import { StudentsRepostiory } from '../repositories/student-repository'
import { HashGenerator } from '../cryptography/hash-generator'
import { StudentAlreadyExitsError } from './errors/student-already-exists-error'
import { Student } from '../../enterprise/entities/student'

interface RegisterStudentUseCaseRequest {
  name: string
  email: string
  password: string
}

type RegisterStudentUseCaseResponse = Either<
  StudentAlreadyExitsError,
  {
    student: Student
  }
>

@Injectable()
export class RegisterStudentUseCase {
  constructor(
    private studentsRepository: StudentsRepostiory,
    private hashGenerator: HashGenerator,
  ) {}

  async execute({
    email,
    name,
    password,
  }: RegisterStudentUseCaseRequest): Promise<RegisterStudentUseCaseResponse> {
    const studentWithSameEmail =
      await this.studentsRepository.findByEmail(email)

    if (studentWithSameEmail) {
      return left(new StudentAlreadyExitsError(email))
    }

    const hashedPassword = await this.hashGenerator.hash(password)

    const student = Student.create({
      email,
      name,
      password: hashedPassword,
    })

    await this.studentsRepository.create(student)

    return right({
      student,
    })
  }
}
